requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scipy>=1.10.0
kagglehub>=0.2.5
python-dotenv>=1.0.0
seaborn>=0.13.0

